requirements.txt
pyyaml
pandas
sqlalchemy
mysql-connector-python